requirements.txt
Flask
pymongo
watson-developer-cloud>=1.5.0
flasgger
bcrypt
python-dotenv
flask-cors
neo4j-driver